# Conflicts:
#	app/Eco/AddressEnergySupplier/AddressEnergySupplierObserver.php

+++ app/Helpers/Delete/Models/DeleteFinancialOverviewParticipantProject.php
@@ -46,13 +46,18 @@ class DeleteFinancialOverviewParticipantProject implements DeleteInterface
      */
     public function delete()
     {
-        $this->canDelete();
+        if (! $this->canDelete()) {
+            return $this->errorMessage;
+        }
+
         $this->deleteModels();
         $this->dissociateRelations();
         $this->deleteRelations();
         $this->customDeleteActions();
-        $this->financialOverviewParticipantProject->delete();
 
+        if (count($this->errorMessage) === 0) {
+            $this->financialOverviewParticipantProject->delete();
+        }
         return $this->errorMessage;
     }
 
@@ -60,6 +65,12 @@ class DeleteFinancialOverviewParticipantProject implements DeleteInterface
      */
     public function canDelete()
     {
+        $isDraft = $this->financialOverviewParticipantProject->status_id === 'concept';
+
+        if ($isDraft) {
+            return true;
+        }
+
         if($this->financialOverviewParticipantProject->status_id === 'sent'){
             array_push($this->errorMessage, "Er zijn al waardestaten voor deelnemer verzonden.");
         }
@@ -75,6 +86,8 @@ class DeleteFinancialOverviewParticipantProject implements DeleteInterface
         if($hasFinancialOverviewContactSent){
             array_push($this->errorMessage, "Er zijn al waardestaten voor contact verzonden.");
         }
+
+        return count($this->errorMessage) === 0;
     }
 
     /** Deletes models recursive
